+++ src/parser/NHXannotation.h
@@ -36,16 +36,16 @@ extern "C" {
    This is how long tag strings (such as 'AC' or 'D') in
    NHX trees are allowed to be, *including* '\0'!
 */
+
 #define MAX_ANNOTATION_ID_LENGTH 5
 
 struct NHXannotation {
     char anno_type[MAX_ANNOTATION_ID_LENGTH];
     union {
-        float t;			/* For node times */
-        unsigned i;			/* Anti-chain (i.e., number of species node) */
-        char *str;			/* For species name, et.c. */
+        double t;               /* For node times */
+        int i;                  /* Anti-chain (i.e., number of species node) */
+        char *str;              /* For species name, et.c. */
         struct int_list *il;	/* List of integers */
-
         /* Preparing for other annotation here! */
     } arg;
 
@@ -54,7 +54,7 @@ struct NHXannotation {
 
 /* Structure for integer lists */
 struct int_list {
-    unsigned i;
+    int i;
     struct int_list *next;
 };
 
@@ -69,20 +69,21 @@ unsigned annotation_isa(struct NHXannotation* l, const char *tag);
 /*
   Prepend a new annotation to the annotation list l:
 */
-struct NHXannotation *new_newick_weight(float t, struct NHXannotation *l);
+struct NHXannotation *new_newick_weight(double t, struct NHXannotation *l);
 struct NHXannotation *new_annotation(char *tag, struct NHXannotation *l);
 struct NHXannotation *new_duplication(struct NHXannotation *l);
-struct NHXannotation *new_anti_chain(unsigned i, struct NHXannotation *l);
+struct NHXannotation *new_anti_chain(int i, struct NHXannotation *l);
 struct NHXannotation *new_species_name(char *str, struct NHXannotation *l);
-struct NHXannotation *new_node_id(unsigned id, struct NHXannotation *l);
+struct NHXannotation *new_node_id(int id, struct NHXannotation *l);
 struct NHXannotation *append_annotations(struct NHXannotation* l1, struct NHXannotation *l2);
 
 /*
   Handle integer lists.
 */
-struct int_list* new_int_list(unsigned i, struct int_list *next);
+struct int_list* new_int_list(int i, struct int_list *next);
 void free_int_list(struct int_list *il);
 struct int_list* int_list_reverse(struct int_list *il);
+
 #ifdef __cplusplus
 }
 #endif
